scripts/requirements.txt
<<<<<<< Updated upstream
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.4
click-log==0.4.0
colorama==0.4.6
courlan==0.9.4
dateparser==1.1.8
deprecation==2.1.0
docutils==0.20.1
dotty-dict==1.3.1
exceptiongroup==1.1.2
feedparser==6.0.10
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.2
h11==0.14.0
htmldate==1.5.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.7.0
invoke==1.7.3
jaraco.classes==3.2.3
jusText==3.0.0
keyring==24.2.0
langcodes==3.3.0
lxml==4.9.2
more-itertools==9.1.0
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
pkginfo==1.9.6
postgrest==0.10.6
pydantic==1.10.11
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.15.0
python-semantic-release==7.33.2
pytz==2023.3.post1
readme-renderer==40.0
realtime==1.0.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
storage3==0.5.2
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tiktoken==0.4.0
tld==0.13
tomlkit==0.11.8
tqdm==4.65.0
trafilatura==1.6.2
tweepy==4.14.0
twine==3.8.0
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.3
webencodings==0.5.1
websockets==10.4
yarl==1.9.2
zipp==3.15.0
=======
feedparser==6.0.11
openai==1.12.0
pip-chill==1.0.3
python-dotenv==1.0.1
supabase==2.3.6
tiktoken==0.6.0
trafilatura==1.7.0
>>>>>>> Stashed changes
